fix(locator) + test(fixture): ordinal width + 2-locator + 3-server cluster ThinClientLocatorHelper.BuildRequestFrame was writing the protocol ordinal as int32; the Java locator's TcpServer.processOneConnection reads it via input.readShort() at TcpServer.java:413, so our int32 left the trailing two bytes mis-aligning the DSCode envelope and the server rejected every locator request with "UnsupportedSerializationVersionException: ordinal 0 not supported". Switched to WriteInt16 (matches Java TcpClient.java:312); cppcache also writes a wider int in its bytestream but happens to not surface this because cppcache's locator round-trip is exercised by other servers, not the test path. Latent because no integration test actually completed a locator-discovered round trip — the locator round-trip test was Skipped for an unrelated NAT reason.
Integration fixture rebuilt to bring up 2 locators + 3 servers in one
container so locator/pool/failover code paths exercise a realistic
topology instead of a single-process degenerate case:
- Locators on container ports 10334-10335, servers on 40404-40406;
  fixed host-port bindings (matching container ports) so each
  server's --hostname-for-clients=localhost resolves from the test
  host. Trade-off: at most one fixture per host (xUnit
  ICollectionFixture already enforces this).
- Wait strategy waits on all 3 server ports before tests run.
- LocatorPort2 / ServerPort2 / ServerPort3 / LocatorPorts /
  ServerPorts / LocatorEndpoints exposed for multi-endpoint tests;
  LocatorPort / ServerPort kept pointing at loc1 / srv1 so the 100+
  existing direct-server tests need no changes.

GeodeContainerSmokeTests gains ClusterReportsTwoLocatorsAndThreeServers
which calls `gfsh list members` and asserts all 5 names appear — a
silently-failed member start would otherwise only surface as strange
downstream failures.

LocatorModeIntegrationTests.Pool_with_locator_supports_region_put_get_round_trip
unskipped; the fixture's hostname-for-clients + fixed-port mappings
now resolve the locator-returned server address from the test host,
and the ordinal-width fix is what made the locator call succeed at
all. Debug logs confirm the locator round-robins across all three
servers on consecutive Put/Get calls.

Tests: 101/106 integration green (5 expected skips, no regressions);
not run on unit suite.

+++ src/Geode.Client/Internal/ThinClientLocatorHelper.cs
@@ -251,7 +251,13 @@ internal sealed class ThinClientLocatorHelper(
         var writer = new BigEndianBinaryWriter(bufferWriter);
 
         writer.WriteInt32(GossipVersion);
-        writer.WriteInt32(ProtocolVersion.Current.Ordinal);
+        // Ordinal MUST be int16 — Java TcpServer.processOneConnection reads
+        // `input.readShort()` at TcpServer.java:413; an int32 here leaves
+        // the trailing 2 bytes mis-aligning the DSCode/DSFid envelope and
+        // the server rejects with
+        // `UnsupportedSerializationVersionException: ordinal 0 not supported`.
+        // Matches Java TcpClient.java:312 (`writeShort(ordinalVersion)`).
+        writer.WriteInt16(ProtocolVersion.Current.Ordinal);
         writer.WriteByte(DSCode.FixedIDByte);
         writer.WriteSByte((sbyte)dsfid);
         writeBody(writer);
